map_app/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cvzone==1.6.1
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
